Addition of Gyro Sensor
Added code to implement the MPU6050 Gyro for Yaw Axis control using a simple function in robot.cpp.
I wrote a test for it in the Main file and changed the robot setup.
In my (very simple, no robot involved) tests for 90° turns this showed a ~2° accuracy.
I add this now although I heard in the milestone meeting we won't want to use the gyro, but I think it could be useful, at least to give it a try.