Addition of i2C Scanner and change of TurnDirGyro Method
I guess the problem yesterday where we weren't able to connect the ToFs was for one a hardware problem as the Gyro wasn't plugged in correctly so the ESP would just always print the i2C errors of no connect.
Fixing this alone first didn't work out as well, but after adding some stuff it now works. During the startup procedure the robot often prints the current state of the i2C bus via the i2CScanner method for debugging purposes.

Then there was an issue with the gyro that could have been the reason for bad performance in recent tests.

Now there's also a corrected turn method that should be more exact. This corrects its turns by turning back and forth multiple times (max: 5 currently). Because it would create complexity and delay if we were to always turn back and forth all the time it also reiterates its brake values so the next turn should always be better than the last from the start (with less correction).
However in my testing the robot may come in a state where the correction doesn't work anymore, but I don't know.